requirements.txt
playwright==1.49.0
anthropic==0.40.0
rich==13.9.0
python-dotenv==1.0.1
pyyaml==6.0.2
